Cleanup Snapshot ITs Further (#63985)
Drying up a few more spots and extracting tests that use their own custom plugins to separate
classes to make the giant `DedicatedClusterSnapshotRestoreIT` suite a little more digestible.
