+++ server/src/internalClusterTest/java/org/elasticsearch/snapshots/RestoreSnapshotIT.java
@@ -20,7 +20,6 @@
 package org.elasticsearch.snapshots;
 
 import org.elasticsearch.action.ActionFuture;
-import org.elasticsearch.action.admin.cluster.snapshots.create.CreateSnapshotResponse;
 import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotResponse;
 import org.elasticsearch.action.admin.indices.settings.get.GetSettingsResponse;
 import org.elasticsearch.action.admin.indices.template.get.GetIndexTemplatesResponse;
@@ -38,6 +37,7 @@ import org.elasticsearch.repositories.RepositoriesService;
 import org.elasticsearch.rest.RestStatus;
 
 import java.nio.file.Path;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.Locale;
@@ -90,24 +90,8 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
 
         createRepository(repoName, "fs", absolutePath);
 
-        logger.info("--> snapshot");
-        CreateSnapshotResponse createSnapshotResponse = client.admin().cluster().prepareCreateSnapshot(repoName, snapshotName1)
-                .setWaitForCompletion(true)
-                .setIndices(indexName1)
-                .get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
-        assertThat(createSnapshotResponse.getSnapshotInfo().state(), equalTo(SnapshotState.SUCCESS));
-
-        CreateSnapshotResponse createSnapshotResponse2 = client.admin().cluster().prepareCreateSnapshot(repoName, snapshotName2)
-                .setWaitForCompletion(true)
-                .setIndices(indexName2)
-                .get();
-        assertThat(createSnapshotResponse2.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse2.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse2.getSnapshotInfo().totalShards()));
-        assertThat(createSnapshotResponse2.getSnapshotInfo().state(), equalTo(SnapshotState.SUCCESS));
+        createSnapshot(repoName, snapshotName1, Collections.singletonList(indexName1));
+        createSnapshot(repoName, snapshotName2, Collections.singletonList(indexName2));
 
         RestoreSnapshotResponse restoreSnapshotResponse1 = client.admin().cluster().prepareRestoreSnapshot(repoName, snapshotName1)
                 .setWaitForCompletion(false)
@@ -147,15 +131,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
 
         createRepository(repoName, "fs", absolutePath);
 
-        logger.info("--> snapshot");
-        CreateSnapshotResponse createSnapshotResponse = client.admin().cluster().prepareCreateSnapshot(repoName, snapshotName)
-                .setWaitForCompletion(true)
-                .setIndices(indexName1, indexName2)
-                .get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
-        assertThat(createSnapshotResponse.getSnapshotInfo().state(), equalTo(SnapshotState.SUCCESS));
+        createSnapshot(repoName, snapshotName, Arrays.asList(indexName1, indexName2));
 
         ActionFuture<RestoreSnapshotResponse> restoreSnapshotResponse1 = client.admin().cluster()
                 .prepareRestoreSnapshot(repoName, snapshotName)
@@ -200,10 +176,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
                 .boxed().collect(Collectors.toMap(shardId -> shardId, indexMetadata::primaryTerm));
 
         createRepository("test-repo", "fs");
-        final CreateSnapshotResponse createSnapshotResponse = clusterAdmin().prepareCreateSnapshot("test-repo", "test-snap")
-                .setWaitForCompletion(true).setIndices(indexName).get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), equalTo(numPrimaries));
-        assertThat(createSnapshotResponse.getSnapshotInfo().failedShards(), equalTo(0));
+        createSnapshot("test-repo", "test-snap", Collections.singletonList(indexName));
 
         assertAcked(client().admin().indices().prepareClose(indexName));
 
@@ -232,12 +205,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         assertAcked(client().admin().indices().preparePutMapping("test-idx").setType("_doc").setSource("baz", "type=text"));
         ensureGreen();
 
-        logger.info("--> snapshot it");
-        CreateSnapshotResponse createSnapshotResponse = clusterAdmin().prepareCreateSnapshot("test-repo", "test-snap")
-                .setWaitForCompletion(true).setIndices("test-idx").get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
+        createSnapshot("test-repo", "test-snap", Collections.singletonList("test-idx"));
 
         logger.info("--> delete the index and recreate it with foo field");
         cluster().wipeIndices("test-idx");
@@ -283,10 +251,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
 
         assertFalse(client().admin().indices().prepareGetAliases("alias-123").get().getAliases().isEmpty());
 
-        logger.info("--> snapshot");
-        assertThat(clusterAdmin().prepareCreateSnapshot("test-repo", "test-snap")
-                        .setIndices().setWaitForCompletion(true).get().getSnapshotInfo().state(),
-                equalTo(SnapshotState.SUCCESS));
+        createSnapshot("test-repo", "test-snap", Collections.emptyList());
 
         logger.info("-->  delete all indices");
         cluster().wipeIndices("test-idx-1", "test-idx-2", "test-idx-3");
@@ -351,11 +316,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
                         .endObject())
                 .get().isAcknowledged(), equalTo(true));
 
-        logger.info("--> snapshot");
-        final SnapshotInfo snapshotInfo = assertSuccessful(clusterAdmin().prepareCreateSnapshot("test-repo", "test-snap")
-                .setIndices().setWaitForCompletion(true).execute());
-        assertThat(snapshotInfo.totalShards(), equalTo(0));
-        assertThat(snapshotInfo.successfulShards(), equalTo(0));
+        createSnapshot("test-repo", "test-snap", Collections.emptyList());
         assertThat(getSnapshot("test-repo", "test-snap").state(), equalTo(SnapshotState.SUCCESS));
 
         logger.info("-->  delete test template");
@@ -374,7 +335,6 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         assertIndexTemplateExists(getIndexTemplatesResponse, "test-template");
     }
 
-
     public void testRenameOnRestore() throws Exception {
         Client client = client();
 
@@ -392,12 +352,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         indexRandomDocs("test-idx-1", 100);
         indexRandomDocs("test-idx-2", 100);
 
-        logger.info("--> snapshot");
-        CreateSnapshotResponse createSnapshotResponse = client.admin().cluster().prepareCreateSnapshot("test-repo", "test-snap")
-                .setWaitForCompletion(true).setIndices("test-idx-1", "test-idx-2").get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
+        createSnapshot("test-repo", "test-snap", Arrays.asList("test-idx-1", "test-idx-2"));
 
         logger.info("--> restore indices with different names");
         RestoreSnapshotResponse restoreSnapshotResponse = client.admin().cluster().prepareRestoreSnapshot("test-repo", "test-snap")
@@ -498,10 +453,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
                 .put("chunk_size", 100, ByteSizeUnit.BYTES));
 
         createIndexWithRandomDocs("test-idx", 100);
-
-        logger.info("--> snapshot");
-        client.admin().cluster().prepareCreateSnapshot("test-repo", "test-snap")
-                .setWaitForCompletion(true).setIndices("test-idx").get();
+        createSnapshot("test-repo", "test-snap", Collections.singletonList("test-idx"));
 
         logger.info("--> delete index");
         cluster().wipeIndices("test-idx");
@@ -562,12 +514,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
         assertHitCount(client.prepareSearch("test-idx").setSize(0).setQuery(matchQuery("field1", "Foo")).get(), 0);
         assertHitCount(client.prepareSearch("test-idx").setSize(0).setQuery(matchQuery("field1", "bar")).get(), numdocs);
 
-        logger.info("--> snapshot it");
-        CreateSnapshotResponse createSnapshotResponse = client.admin().cluster().prepareCreateSnapshot("test-repo", "test-snap")
-                .setWaitForCompletion(true).setIndices("test-idx").get();
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-        assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
+        createSnapshot("test-repo", "test-snap", Collections.singletonList("test-idx"));
 
         logger.info("--> delete the index and recreate it while changing refresh interval and analyzer");
         cluster().wipeIndices("test-idx");
@@ -667,12 +614,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
                 client().admin().indices().prepareUpdateSettings("test-idx").setSettings(initialSettingsBuilder).get();
             }
 
-            logger.info("--> snapshot index");
-            CreateSnapshotResponse createSnapshotResponse = client.admin().cluster().prepareCreateSnapshot("test-repo", "test-snap")
-                    .setWaitForCompletion(true).setIndices("test-idx").get();
-            assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(), greaterThan(0));
-            assertThat(createSnapshotResponse.getSnapshotInfo().successfulShards(),
-                    equalTo(createSnapshotResponse.getSnapshotInfo().totalShards()));
+            createSnapshot("test-repo", "test-snap", Collections.singletonList("test-idx"));
 
             logger.info("--> remove blocks and delete index");
             disableIndexBlock("test-idx", IndexMetadata.SETTING_BLOCKS_METADATA);
@@ -736,8 +678,7 @@ public class RestoreSnapshotIT extends AbstractSnapshotIntegTestCase {
             indexRandomDocs("test-index", between(0, 100));
             flush("test-index");
         }
-        clusterAdmin().prepareCreateSnapshot("test-repo", "snapshot-0")
-                .setIndices("test-index").setWaitForCompletion(true).get();
+        createSnapshot("test-repo", "snapshot-0", Collections.singletonList("test-index"));
         final SnapshotRestoreException restoreError = expectThrows(SnapshotRestoreException.class,
                 () -> clusterAdmin().prepareRestoreSnapshot("test-repo", "snapshot-0")
                         .setIndexSettings(Settings.builder().put(INDEX_SOFT_DELETES_SETTING.getKey(), false))
